+++ CHANGELOG.md
@@ -63,6 +63,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Changed background snapshotting to suppress stdout and stderr to avoid long
   hangs.
 
+* `jj gerrit upload` now supports a variety of new flags documented in
+  [gerrit's documentation](https://gerrit-review.googlesource.com/Documentation/user-upload.html).
+  This includes, for example, `--reviewer=foo@example.com` and
+  `--label=Auto-Submit`.
+
 ### Fixed bugs
 
 ## [0.38.0] - 2026-02-04